+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Unreleased
 ### Added
+- Support for fetching modules
 - Support for committing modules